+++ server/service/handler.go
@@ -31,7 +31,6 @@ type FleetEndpoints struct {
 	CreateUserWithInvite         endpoint.Endpoint
 	PerformRequiredPasswordReset endpoint.Endpoint
 	VerifyInvite                 endpoint.Endpoint
-	CarveBlock                   endpoint.Endpoint
 	InitiateSSO                  endpoint.Endpoint
 	CallbackSSO                  endpoint.Endpoint
 	SSOSettings                  endpoint.Endpoint
@@ -61,11 +60,6 @@ func MakeFleetServerEndpoints(svc fleet.Service, urlPrefix string, limitStore th
 		// PerformRequiredPasswordReset needs only to authenticate the
 		// logged in user
 		PerformRequiredPasswordReset: logged(canPerformPasswordReset(makePerformRequiredPasswordResetEndpoint(svc))),
-
-		// For some reason osquery does not provide a node key with the block
-		// data. Instead the carve session ID should be verified in the service
-		// method.
-		CarveBlock: logged(makeCarveBlockEndpoint(svc)),
 	}
 }
 
@@ -77,7 +71,6 @@ type fleetHandlers struct {
 	CreateUserWithInvite         http.Handler
 	PerformRequiredPasswordReset http.Handler
 	VerifyInvite                 http.Handler
-	CarveBlock                   http.Handler
 	InitiateSSO                  http.Handler
 	CallbackSSO                  http.Handler
 	SettingsSSO                  http.Handler
@@ -96,7 +89,6 @@ func makeKitHandlers(e FleetEndpoints, opts []kithttp.ServerOption) *fleetHandle
 		CreateUserWithInvite:         newServer(e.CreateUserWithInvite, decodeCreateUserRequest),
 		PerformRequiredPasswordReset: newServer(e.PerformRequiredPasswordReset, decodePerformRequiredPasswordResetRequest),
 		VerifyInvite:                 newServer(e.VerifyInvite, decodeVerifyInviteRequest),
-		CarveBlock:                   newServer(e.CarveBlock, decodeCarveBlockRequest),
 		InitiateSSO:                  newServer(e.InitiateSSO, decodeInitiateSSORequest),
 		CallbackSSO:                  newServer(e.CallbackSSO, decodeCallbackSSORequest),
 		SettingsSSO:                  newServer(e.SSOSettings, decodeNoParamsRequest),
@@ -283,7 +275,6 @@ func attachFleetAPIRoutes(r *mux.Router, h *fleetHandlers) {
 	r.Handle("/api/v1/fleet/sso/callback", h.CallbackSSO).Methods("POST").Name("callback_sso")
 	r.Handle("/api/v1/fleet/users", h.CreateUserWithInvite).Methods("POST").Name("create_user_with_invite")
 	r.Handle("/api/v1/fleet/invites/{token}", h.VerifyInvite).Methods("GET").Name("verify_invite")
-	r.Handle("/api/v1/osquery/carve/block", h.CarveBlock).Methods("POST").Name("carve_block")
 }
 
 func attachNewStyleFleetAPIRoutes(r *mux.Router, svc fleet.Service, logger kitlog.Logger, opts []kithttp.ServerOption) {
@@ -442,6 +433,10 @@ func attachNewStyleFleetAPIRoutes(r *mux.Router, svc fleet.Service, logger kitlo
 	// with the request.
 	ne := newNoAuthEndpointer(svc, opts, r, "v1")
 	ne.POST("/api/_version_/osquery/enroll", enrollAgentEndpoint, enrollAgentRequest{})
+
+	// For some reason osquery does not provide a node key with the block data.
+	// Instead the carve session ID should be verified in the service method.
+	ne.POST("/api/_version_/osquery/carve/block", carveBlockEndpoint, carveBlockRequest{})
 }
 
 // TODO: this duplicates the one in makeKitHandler
