--- server/service/service_carves.go
@@ -1,46 +1,0 @@
-package service
-
-import (
-	"context"
-	"errors"
-	"fmt"
-
-	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
-	"github.com/fleetdm/fleet/v4/server/fleet"
-)
-
-func (svc *Service) CarveBlock(ctx context.Context, payload fleet.CarveBlockPayload) error {
-	// skipauth: Authorization is currently for user endpoints only.
-	svc.authz.SkipAuthorization(ctx)
-
-	// Note host did not authenticate via node key. We need to authenticate them
-	// by the session ID and request ID
-	carve, err := svc.carveStore.CarveBySessionId(ctx, payload.SessionId)
-	if err != nil {
-		return ctxerr.Wrap(ctx, err, "find carve by session_id")
-	}
-
-	if payload.RequestId != carve.RequestId {
-		return errors.New("request_id does not match")
-	}
-
-	// Request is now authenticated
-
-	if payload.BlockId > carve.BlockCount-1 {
-		return fmt.Errorf("block_id exceeds expected max (%d): %d", carve.BlockCount-1, payload.BlockId)
-	}
-
-	if payload.BlockId != carve.MaxBlock+1 {
-		return fmt.Errorf("block_id does not match expected block (%d): %d", carve.MaxBlock+1, payload.BlockId)
-	}
-
-	if int64(len(payload.Data)) > carve.BlockSize {
-		return fmt.Errorf("exceeded declared block size %d: %d", carve.BlockSize, len(payload.Data))
-	}
-
-	if err := svc.carveStore.NewBlock(ctx, carve, payload.BlockId, payload.Data); err != nil {
-		return ctxerr.Wrap(ctx, err, "save block data")
-	}
-
-	return nil
-}
